Day 20: solution to hard problem w/o additional assumptions
The problem of day 20 is posed in a general way which is much harder to solve
compared to what is actually given as input. In particular, the input seems to
guarantee that each an every branch (regex group alternative) returns to the
same point, e.g. ^N(NESW|)N$ while branches ending up in different locations
such as ^N(W|E)N$ are apparently disallowed. As a consequence, consecutive
branches such as ^N(WWWWW|EEEEE)N(WWW|EEE)N(W|E)N$ which would multiply the
number of possible total branches (2*2*2=8) are not possible; instead, the
number of separate branches that need to be considered is only additive, e.g.
^N(NWWWWWS|WWWWW)N(NWWWS|WWW)N(NWS|W)N$ results in 2+2+2=6 branches.

While the present program *does* consider all possible paths (w/o the above
assumption) and produces the correct solution for the small examples given on
the website, the exponential growth of the number of total branches is
prohibitively expensive to process the full puzzle input in finite time.